refactor(email_service/send_email): ensure ses.py contains only necessary SES code (SCRUM-73)
