+++ src/email_service/send_email/ses.py
@@ -1,40 +1,18 @@
 import logging
-import os
-import re
 from email.mime.application import MIMEApplication
 from pathlib import Path
 
 import boto3
-import requests
 import time_util
 from botocore.exceptions import ClientError
 from certificate_generator import generate_certificate
-from current_user_util import current_user_util
-from email_repository import EmailRepository
 from email_util import attach_files_to_message, create_email_message
-from file_util import download_file_content
-from s3 import read_html_template_file_from_s3
 from template_util import replace_placeholders
-
-from file_service import FileService
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 ses_client = boto3.client("ses", region_name="ap-northeast-1")
-
-CHARSET = "utf-8"
-BUCKET_NAME = os.getenv("BUCKET_NAME")
-PRIVATE_BUCKET_NAME = os.getenv("PRIVATE_BUCKET_NAME")
-CERTIFICATE_TEMPLATE_FILE_S3_OBJECT_KEY = (
-    "templates/[template] AWS Educate certificate.pdf"
-)
-
-# Initialize FileService
-file_service = FileService()
-
-# Initialize EmailRepository
-email_repository = EmailRepository()
 
 
 def send_email(
@@ -132,54 +110,3 @@ def send_email(
             logger.error("Failed to delete certificate file: %s", e)
 
     return None, "FAILED"
-
-
-def process_email(
-    email_data: dict,
-    row: list[dict[str, any]],
-) -> tuple:
-    """
-    Process email sending for a given row of data.
-
-    :param email_data: Dictionary containing email metadata such as subject, display_name, template_file_id, etc.
-    :param row: A dictionary representing a single row of data from Excel, containing email recipient and placeholder values.
-    :return: Status and email ID of the email sending operation.
-    """
-    recipient_email = row.get("Email")
-    email_id = email_data.get("email_id")
-
-    logger.info("Row data: %s", row)
-    if not re.match(r"[^@]+@[^@]+\.[^@]+", recipient_email):
-        logger.warning("Invalid email address provided: %s", recipient_email)
-        return "FAILED", email_id
-
-    # Get template file information and content
-    template_file_info = file_service.get_file_info(
-        email_data.get("template_file_id"),
-        current_user_util.get_current_user_access_token(),
-    )
-    template_file_s3_object_key = template_file_info["s3_object_key"]
-    template_content = read_html_template_file_from_s3(
-        bucket=BUCKET_NAME, template_file_s3_key=template_file_s3_object_key
-    )
-
-    # Send email
-    _, status = send_email(
-        email_data.get("subject"),
-        template_content,
-        row,
-        email_data.get("display_name"),
-        email_data.get("reply_to"),
-        email_data.get("sender_local_part"),
-        email_data.get("attachment_file_ids"),
-        email_data.get("is_generate_certificate"),
-        email_data.get("run_id"),
-        email_data.get("cc"),
-        email_data.get("bcc"),
-    )
-
-    # Update the item in DynamoDB using EmailRepository
-    email_repository.update_email_status(
-        run_id=email_data.get("run_id"), email_id=email_id, status=status
-    )
-    return status, email_id
