+++ src/email_service/send_email/template_util.py
@@ -1,4 +1,3 @@
-# template_util.py
 import re
